requirements.txt
django>=5.2
djangorestframework
paddleocr
paddlepaddle
opencv-python
numpy
